# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.Category:
    properties:
      name:
        type: string
    type: object
...
  model.Info:
    properties:
      desc:
        type: string
    type: object
...
  request.PageInfo:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  license: {}
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /category/createCategory:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建Category
      tags:
      - Category
  /category/deleteCategory:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除Category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Category
      tags:
      - Category
  /category/findCategory:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询Category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询Category
      tags:
      - Category
  /category/getCategoryList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取Category列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取Category列表
      tags:
      - Category
  /category/updateCategory:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新Category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Category
      tags:
      - Category
...
  /info/createInfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建Info
      tags:
      - Info
  /info/deleteInfo:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Info
      tags:
      - Info
  /info/findInfo:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询Info
      tags:
      - Info
  /info/getInfoList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取Info列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取Info列表
      tags:
      - Info
  /info/updateInfo:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新Info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Info'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Info
      tags:
      - Info
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"